backend/requirements.txt
flask==3.0.2
pytesseract==0.3.10
pillow==10.1.0
python-dotenv==1.0.0
flask-cors==4.0.0
pdf2image==1.17.0
werkzeug==3.0.1